refactor(coalescent): extract coalescent/ from commands/timetree/ to top-level module
- Move 13 source files, tests, and fixtures to top-level src/coalescent/
- Rewrite all import paths and fixture directory constants
- Update example and capture script paths
