+++ packages/treetime/src/coalescent/__tests__/test_events.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod tests {
   use crate::clock::date_constraints::load_date_constraints;
-  use crate::commands::timetree::coalescent::events::collect_tree_events;
-  use crate::commands::timetree::coalescent::time_coordinate::CalendarTime;
+  use crate::coalescent::events::collect_tree_events;
+  use crate::coalescent::time_coordinate::CalendarTime;
   use crate::pretty_assert_ulps_eq;
   use crate::representation::partition::timetree::GraphTimetree;
   use eyre::Report;
