+++ packages/treetime/src/coalescent/__tests__/test_gm_total_lh.rs
@@ -14,7 +14,7 @@
 mod tests {
   use super::super::helpers::setup_graph;
   use crate::clock::date_constraints::load_date_constraints;
-  use crate::commands::timetree::coalescent::total_lh::compute_coalescent_total_lh;
+  use crate::coalescent::total_lh::compute_coalescent_total_lh;
   use crate::representation::partition::timetree::GraphTimetree;
   use approx::assert_abs_diff_eq;
   use eyre::Report;
