+++ packages/treetime/src/coalescent/edge_data.rs
@@ -1,6 +1,6 @@
-use crate::commands::timetree::coalescent::piecewise_constant_fn::PiecewiseConstantFn;
-use crate::commands::timetree::coalescent::piecewise_linear_fn::PiecewiseLinearFn;
-use crate::commands::timetree::coalescent::time_coordinate::{CalendarTime, Tbp};
+use crate::coalescent::piecewise_constant_fn::PiecewiseConstantFn;
+use crate::coalescent::piecewise_linear_fn::PiecewiseLinearFn;
+use crate::coalescent::time_coordinate::{CalendarTime, Tbp};
 use crate::representation::payload::traits::TimetreeNode;
 use eyre::Report;
 use log::warn;
