+++ packages/treetime/src/coalescent/piecewise_constant_fn.rs
