+++ packages/treetime/src/coalescent/total_lh.rs
@@ -1,7 +1,7 @@
-use crate::commands::timetree::coalescent::edge_data::{collect_coalescent_edges, sum_coalescent_cost};
-use crate::commands::timetree::coalescent::events::collect_tree_events;
-use crate::commands::timetree::coalescent::integration::compute_integral_merger_rate;
-use crate::commands::timetree::coalescent::lineage_dynamics::compute_lineage_count_distribution;
+use crate::coalescent::edge_data::{collect_coalescent_edges, sum_coalescent_cost};
+use crate::coalescent::events::collect_tree_events;
+use crate::coalescent::integration::compute_integral_merger_rate;
+use crate::coalescent::lineage_dynamics::compute_lineage_count_distribution;
 use crate::representation::payload::traits::TimetreeNode;
 use eyre::Report;
 use treetime_distribution::Distribution;
