+++ packages/treetime/src/commands/timetree/mod.rs
@@ -1,5 +1,4 @@
 pub mod args;
-pub mod coalescent;
 pub mod convergence;
 pub mod inference;
 pub mod initialization;
