+++ packages/treetime/src/commands/timetree/run.rs
@@ -5,10 +5,10 @@ use crate::clock::clock_output::write_clock_model;
 use crate::clock::clock_regression::{ClockParams, estimate_clock_model_with_reroot_policy};
 use crate::clock::find_best_root::params::BranchPointOptimizationParams;
 use crate::clock::reroot::RerootParams;
+use crate::coalescent::optimize_tc::optimize_tc;
+use crate::coalescent::skyline::{SkylineParams, optimize_skyline};
 use crate::commands::shared::args::BranchLengthMode;
 use crate::commands::timetree::args::{TimeMarginalMode, TreetimeTimetreeArgs};
-use crate::commands::timetree::coalescent::optimize_tc::optimize_tc;
-use crate::commands::timetree::coalescent::skyline::{SkylineParams, optimize_skyline};
 use crate::commands::timetree::convergence::metrics::{IterationContext, TimetreeOptimizer};
 use crate::commands::timetree::inference::runner::run_timetree;
 use crate::commands::timetree::initialization::{InputData, initialize_partitions, load_input_data};
